Add mock franka subpackage
mock/ creates an implementation of a server and various backend to
emulate a real franka arm, which can be connected to through TCP/UDP
communication, as if connecting to a real arm. Various backends can be
used to compute the next robot state from an input command. Applications
will want to import this package and run the server with their own
backends, or an example can be launched through:

<build_dir>/mock_server_example <127.0.0.1 or other IP>