formatting and linting
Fixes initial joint position assignment

Ensures that the initial joint position is correctly assigned when time is zero in the external control loop example.
This prevents potential issues with rate limiting and trajectory generation in the first iteration.

linting  and formatting fix

Removes RealtimeConfig argument

Removes the `RealtimeConfig::kIgnore` argument from the `franka::Robot` constructor.

This simplifies the example by using the default realtime configuration.

Makes joint number configurable

Defines the number of joints as a constant, allowing for easier
modification and preventing hardcoded values. This change ensures
consistency across the codebase when referring to the number of joints.

Uses kNumJoints for joint velocity limits

Generalizes joint velocity limit functions by using the
kNumJoints constant instead of hardcoding the value 7.

This improves code maintainability and reduces the risk of
inconsistencies if the number of joints changes in the future.